Fix home screen sizing and styling
