--- src/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="./assets/css/daisyui-full.min.css" rel="stylesheet" type="text/css" />
-    <script src="./assets/js/tailwind-play-cdn.js" defer></script>
-    <script src="./assets/js/htmx.min.js" defer></script>
-    <script src="./assets/js/sse.js" defer></script>
-    <script src="./assets/js/htmx-process.js" defer></script>
-    <title>PyHTMX GUI Client</title>
-</head>
-<body>
-    <div id="root" class="flex grow" hx-ext="sse" sse-connect="/updates" sse-swap="root"></div>
-</body>
-</html>
